performed a code clean-up and broke down the routes to make use of controllers
