+++ controllers/profile.js
@@ -1,0 +1,16 @@
+const handleProfileGet = (req, res) => {
+    const { id } = req.params;
+    db.select('*').from('users').where({id})    // used  javascript shorthand object({id: id})
+    .then(user => {
+        if (user.length) {
+            res.json(user[0]);
+        } else {
+            res.status(400).json('Not found');
+        }
+    })
+    .catch(err => res.status(400).json('Error getting user'));
+}
+
+module.exports = {
+    handleProfileGet: handleProfileGet
+}
